Adding some UI and refactor the code
