+++ index.html
@@ -16,7 +16,10 @@
   <body>
     <header class="header">
       <h1>The Game of Life</h1>
-      <!-- <button class="start-btn" onclick="fuck()">START</button> -->
+      
+      <!-- <button class="start-btn" onclick="fiatLife()">START</button> -->
+      <div class="button-start">START</div>
+
       <div class="header-info">
         <div class="header-info__width">
           Width: <span class="header-info__width--value"> </span>
@@ -36,5 +39,18 @@
         <div class="container"></div>
       </div>
     </main>
+
+    <!-- <footer class="footer">
+      <form class="footer-inputs">
+        <ul class="input-form">
+          <li class="input-form__width">width</li>
+          <li class="input-form__height">height</li>
+          <li class="input-form__gameDuration">duration</li>
+          <li class="input-form__renderingSpeed">rendering speed</li>
+        </ul>
+      </form> -->
+    </footer>
+
+
   </body>
 </html>
