+++ src/logic/game-of-life/game-of-life.js
@@ -1,14 +1,15 @@
-import { mapCells } from '../../../helpers/map-cells/map-cells.js';
-import { getNeighbours } from '../../../helpers/get-neighbours/get-neighbours.js';
-import { turnOnCells } from '../../../helpers/turn-on-cells/turn-on-cells.js';
-import { removeProperties } from '../../../helpers/remove-properties/remove-properties.js';
-import { filterAliveCells } from '../../../helpers/filter-alive-cells/filter-alive-cells.js';
-import { consoleRender } from '../../../rendering/console-render/console-render.js';
-import { renderMatrix } from '../../../rendering/render-matrix/render-matrix.js';
-import { matrixValues, isThereLife } from '../../../app.js';
-import { lifeStatus } from '../life-status/life-status.js';
+import { mapCells } from '../../helpers/map-cells/map-cells.js';
+import { getNeighbours } from '../../helpers/get-neighbours/get-neighbours.js';
+import { turnOnCells } from '../../helpers/turn-on-cells/turn-on-cells.js';
+import { removeProperties } from '../../helpers/remove-properties/remove-properties.js';
+import { filterAliveCells } from '../../helpers/filter-alive-cells/filter-alive-cells.js';
+import { consoleRender } from '../../rendering/console-render/console-render.js';
+// import { renderMatrix } from '../../rendering/render-matrix/render-matrix.js';
+import { matrixValues, isThereLife } from '../../app.js';
+import { lifeStatus } from '../../logic/life-status/life-status.js';
+import renderMatrix from '../../rendering/render-matrix/render-matrix.js';
 
-export function gameOfLife(matrix) {
+export default function gameOfLife(matrix) {
   // Step 1: map cells to define their absolute position within the matrix -mutation-
   mapCells(matrix, matrixValues.width, matrixValues.height);
